+++ packages/commonwealth/server/migrations/20241112084149-augment-token-model.js
@@ -25,7 +25,7 @@ module.exports = {
         'Tokens',
         'eth_market_cap_target',
         {
-          type: Sequelize.DECIMAL(78, 0),
+          type: Sequelize.FLOAT,
           allowNull: false,
         },
         { transaction },
